refactor: Namespaces organizados para que classes fiquem agrupadas por estarem relacionadas ao mesmo domínio
