+++ src/Teste/LivroDeReceitas.Api.Test/GlobalUsings.cs
@@ -2,12 +2,10 @@ global using Xunit;
 global using Util.Requests;
 global using AutoMapper;
 global using LivroDeReceitas.Api.Services.Usuarios;
-global using LivroDeReceitas.Domain.Interfaces.Data;
 global using Util.Mapper;
 global using Util.Repositories;
 global using Util.Encryptor;
 global using Util.Token;
 global using LivroDeReceitas.Infrastructure.Validators.Usuarios;
-global using LivroDeReceitas.Domain.DTOs.Usuarios;
 global using LivroDeReceitas.Domain.Exceptions;
 global using LivroDeReceitas.Infrastructure.Services.Usuarios;
